Autelis and Atlona packages renamed to .internal. See openhab/static-code-analysis#100
Also-by: Svilen Valkanov svilen.valkanov@musala.com
