+++ addons/binding/org.openhab.binding.atlona/src/main/java/org/openhab/binding/atlona/internal/discovery/AtlonaDiscovery.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.atlona.discovery;
+package org.openhab.binding.atlona.internal.discovery;
 
 import java.io.IOException;
 import java.net.DatagramPacket;
