+++ addons/binding/org.openhab.binding.atlona/src/main/java/org/openhab/binding/atlona/internal/pro3/AtlonaPro3Config.java
@@ -8,7 +8,7 @@
  */
 package org.openhab.binding.atlona.internal.pro3;
 
-import org.openhab.binding.atlona.discovery.AtlonaDiscovery;
+import org.openhab.binding.atlona.internal.discovery.AtlonaDiscovery;
 
 /**
  * Configuration class for the Atlona Pro3 line of switchers
